+++ functions/src/device/routes/registerDevice.ts
