+++ functions/src/index.ts
@@ -1,1 +1,1 @@
-export * from './device';
+export * from './device/routes/registerDevice';
